MonsterKarte und MonsterkarteImpl in Kartentypen verändert dadurch MonsterCards.java und MonsterCardsImpl.java gelöscht
