+++ app/src/main/java/com/example/munchkin/Karte/KartenTypen/MonsterKarte.java
@@ -1,8 +1,17 @@
 package com.example.munchkin.Karte.KartenTypen;
 
 public interface MonsterKarte extends Türkarte {
-    //int anzahlSchätze;
 
-    int getStufe();
-    void setStufe(int stufe);
+    int getLevel();
+    void setLevel(int level);
+
+    int getAnzahlSchätze();
+    void setAnzahlSchätze(int anzahlSchätze);
+
+    //Übernommen von MonsterCards
+    void badStuff(int playerLevel, int badCategory, int levelMinus);
+    void runAway();
+    int getBadCategory();
+    void setBadCategory(int badCategory);
+
 }
