--- app/src/main/java/com/example/munchkin/Karte/MonsterCards.java
@@ -1,13 +1,0 @@
-package com.example.munchkin.Karte;
-
-public interface MonsterCards  extends Karte{
-
-     int getLevel();
-     void setLevel(int level);
-     void badStuff(int playerLevel, int badCategory, int levelMinus);
-     void runAway();
-     int getBadCategory();
-     void setBadCategory(int badCategory);
-
-
-}
